add employee data validation

+++ src/controller/employee.controller.ts
@@ -4,6 +4,7 @@ import xlsx from "xlsx";
 import * as fs from "fs";
 import path from "path";
 import employee from "../model/employee.model";
+import multer from 'multer';
 
 interface EmpData{
   "No": number,
@@ -27,7 +28,7 @@ interface EmpData{
   "Tanggal Lahir": string,
   "Usia (th.)": number,
   "Pendidikan Terakhir": string,
-  "Pendidikan Terakhir (short)": string,
+  "Pendidikan Terakhir (Short)": string,
   "Agama": string,
   "Jabatan": string,
   "Kode Unit": string,
@@ -46,7 +47,7 @@ interface EmpData{
   "Masa kerja seluruhnya Bln": number,
   "HP": string,
   "Email": string
-}
+};
 
 
 const getAllEmployee = async(req: Request, res: Response, next: NextFunction) => {
@@ -61,14 +62,24 @@ const getAllEmployee = async(req: Request, res: Response, next: NextFunction) =>
 const injectEmployee = async(req: Request, res: Response, next: NextFunction) => {
   const client= await pool.connect();
   try{
-    const workbook = xlsx.readFile(path.join(__dirname, `/Data_pegawai_Mei_2024.xlsx`));
+    if(!req.file){
+      return res.status(400).json({ success: false, message: 'No file uploaded' });
+    };
+
+    const workbook = xlsx.read(req.file.buffer, { type: 'buffer' });
     const jsonXl: EmpData[] = xlsx.utils.sheet_to_json(workbook.Sheets.Sheet1, {defval: null});
+
+    if (!jsonXl.every(validateEmpData)) {
+      return res.status(400).json({ success: false, message: 'Invalid file format' });
+    };
 
     await client.query("BEGIN");
     await employee.deleteAllEmployee(client);
+
     let result = jsonXl.map((item: EmpData) => {
       employee.addEmployee(client, item);
     });
+
     await client.query("COMMIT");
 
     res.status(200).json({sucess: true, message: 'Add employee success', rows: result});
@@ -94,3 +105,20 @@ const deleteAllEmployee = async(req: Request, res: Response, next: NextFunction)
 
 
 export {getAllEmployee, injectEmployee, deleteAllEmployee}
+
+
+// --------------------------------------------------------------------------------------------------------
+function validateEmpData (data: any): data is EmpData {
+  const requiredFields: (keyof EmpData)[] = [
+    "No", "NIP Format Baru", "NIP", "Nama Gelar", "Gelar Depan", "Nama", "Gelar Belakang",
+    "Nama Terakhir", "Sex", "BMI", "Ket BMI", "Pangkat", "Gol", "Gol TMT", "Eselon",
+    "Eselon TMT", "TMT Jabatan", "Tempat Lahir", "Tanggal Lahir", "Usia (th.)",
+    "Pendidikan Terakhir", "Pendidikan Terakhir (Short)", "Agama", "Jabatan", "Kode Unit",
+    "Unit Eselon II", "Unit Eselon III", "Unit Eselon IV", "Unit Kerja", "Unit Sulit Trs",
+    "Remote Zone", "Status", "Aktif Y/N", "Unit Pertama", "Masa kerja golongan Thn",
+    "Masa kerja golongan Bln", "Masa kerja seluruhnya Thn", "Masa kerja seluruhnya Bln",
+    "HP", "Email"
+  ];
+
+  return requiredFields.every(field => field in data);
+};
